Allow several domains in a haproxy configuration file
Additional domains are read from a configuration for each service
(in case the file is present). The path to the file is
`${HAPROXY_OPTIONS_PATH}/<Service ID>_<Service Port>/domains`.
The content of the files is ignored for services that do not use the
HTTP protocol.

General options of a service to pass to haproxy are read from a
configuration file located at
`${HAPROXY_OPTIONS_PATH}/<Service ID>_<Service Port>/config`.